Add package metadata to the CXR Foundation library, to enable easier installation (pip install from git).
Using a modern Hatchling build backend, to keep this in the Python Packaging Authority ecosystem, but avoiding setuptools due to the lack of support for pip hash verification flags in the requirements.txt files. Keeping dependencies dynamic (leveraging the existing requirements.txt files) to avoid duplication and potential out-of-sync issues for the dependencies spec.

PiperOrigin-RevId: 720631222